requirements.txt
Django=>1.9
django-autocomplete-light>=3.1.6
python-memcached

# testes:
Unidecode>=0.04.19
factory_boy>=2.7.0